refactor: Consolidate ProviderError into a single module and remove redundant file
