+++ src/market_data_ingestor/src/providers/mod.rs
@@ -25,9 +25,9 @@
 //! }
 //! ```
 //! 
-pub mod errors;
 
 use async_trait::async_trait;
+use thiserror::Error;
 
 use crate::{errors::Error, models::{bar_series::BarSeries, request_params::BarsRequestParams}};
 
@@ -36,60 +36,22 @@ pub trait DataProvider {
     async fn fetch_bars(&self, params: BarsRequestParams) -> Result<Vec<BarSeries>, Error>;
 }
 
-#[cfg(test)]
-mod tests {
-    use async_trait::async_trait;
-    use chrono::Utc;
+/// Errors that can occur within a `DataProvider` implementation.
+#[derive(Debug, Error)]
+pub enum ProviderError {
+    /// An error during an API request (e.g., network failure, timeout).
+    #[error("API request failed: {0}")]
+    Request(#[from] reqwest::Error),
 
-    use crate::models::{asset::AssetClass, timeframe::{TimeFrame, TimeFrameUnit}};
+    /// The provider's API returned a specific error message (e.g., invalid API key).
+    #[error("API error: {0}")]
+    Api(String),
 
-    use super::*;
-    
-    struct AlpacaProvider;
-    struct PolygonProvider;
+    /// The request parameters were invalid for this specific provider.
+    #[error("Invalid parameters for provider: {0}")]
+    Validation(String),
 
-    #[async_trait]
-    impl DataProvider for AlpacaProvider {
-        async fn fetch_bars(&self, params: BarsRequestParams) -> Result<Vec<BarSeries>, Error> {
-            println!("Fetching from alpaca for symbols: {:?}", params.symbols);
-            Ok(vec![])
-        }
-    }
-
-    #[async_trait]
-    impl DataProvider for PolygonProvider {
-        async fn fetch_bars(&self, params: BarsRequestParams) -> Result<Vec<BarSeries>, Error> {
-            println!("Fetching from Polygon.io for symbols: {:?}", params.symbols);
-            Ok(vec![]) // Return dummy data
-        }
-    }
-
-    // This function decides AT RUNTIME which provider to give back.
-    // It can only work because it returns a `Box<dyn DataProvider>`.
-    fn get_provider(name: &str) -> Box<dyn DataProvider> {
-        if name == "alpaca" {
-            Box::new(AlpacaProvider)
-        } else {
-            Box::new(PolygonProvider)
-        }
-    }
-
-    #[tokio::test]
-    async fn test_dynamic_provider() {
-        // We get a provider, but we don't know or care which one it is.
-        // We just know it implements `DataProvider`.
-        let provider = get_provider("polygon");
-
-        let params = BarsRequestParams {
-            symbols: vec!["ESU24".to_string()],
-            timeframe: TimeFrame::new(1, TimeFrameUnit::Day),
-            start: Utc::now(),
-            end: Utc::now(),
-            asset_class: AssetClass::Futures,
-        };
-
-        // We can call `fetch_bars` because it's part of the trait contract.
-        let result = provider.fetch_bars(params).await;
-        assert!(result.is_ok());
-    }
+    /// An internal error occurred while processing data within the provider.
+    #[error("Internal provider error: {0}")]
+    Internal(String),
 }
